First upload macChanger script
First upload 
Script disconnects wireless network if found to be using default mac address and changes default address to random address, continues to check that the default address is not exposed, if it is, it disconnects from the current wireless network and changes the address with a random one. You do have to reconnect to the wireless network manually every time this happens, this is intentional.